Move pipeline helper functions from modular_pipeline.py
